refactors modules except url_scraper and writer
