+++ lib/delphi/scraper_controller.ex
@@ -8,14 +8,14 @@ defmodule ScraperController do
   end
 
   defp criteria(data) do
-    %Delphi.DataWriter{title: data.title,
+    %Delphi.DataTable{title: data.title,
     description: data.description,
     url: data.url}
     |> clause_match
   end
 
   defp clause_match(data1) do
-    changeset = Delphi.DataWriter.changeset(data1)
+    changeset = Delphi.DataTable.changeset(data1)
     case Delphi.Repo.insert(changeset) do
       {:ok, _} -> IO.puts("======================it worked======================")
       {:error, _} -> IO.puts("failed, possibly unique index...")
